MAT-8675: stop adding data requirements to bundle a second time whhen creating measure bundle
