+++ src/test/java/gov/cms/madie/madiefhirservice/services/HumanReadableServiceTest.java
@@ -48,7 +48,6 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.argThat;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
@@ -65,7 +64,7 @@ class HumanReadableServiceTest
 
   private Library library;
 
-  private org.hl7.fhir.r5.model.Library effectiveDataRequirements;
+  private org.hl7.fhir.r4.model.Library r4EffectiveDataRequirements;
 
   List<MarkdownType> terms = new ArrayList<>();
   MarkdownType term1 = new MarkdownType("Term1 - Definition1");
@@ -108,6 +107,12 @@ class HumanReadableServiceTest
 
     terms.add(term1);
     terms.add(term2);
+
+    r4EffectiveDataRequirements =
+        convertToFhirR4Resource(
+            getStringFromTestResource("/humanReadable/effective" + "-data-requirements.json"),
+            org.hl7.fhir.r4.model.Library.class);
+
     measure =
         new org.hl7.fhir.r4.model.Measure()
             .setName(madieMeasure.getCqlLibraryName())
@@ -123,17 +128,14 @@ class HumanReadableServiceTest
             .setDisclaimer(madieMeasure.getMeasureMetaData().getDisclaimer())
             .setDefinition(terms);
 
+    measure.addContained(r4EffectiveDataRequirements);
+
     String cqlData = ResourceUtils.getData("/test-cql/cv_populations.cql");
     library =
         new Library()
             .addContent(new Attachment().setData(cqlData.getBytes()).setContentType("text/cql"));
 
     library.setId(madieMeasure.getCqlLibraryName());
-
-    effectiveDataRequirements =
-        convertToFhirR5Resource(
-            org.hl7.fhir.r5.model.Library.class,
-            getStringFromTestResource("/humanReadable/effective-data-requirements.json"));
   }
 
   public Bundle.BundleEntryComponent getBundleEntryComponent(Resource resource) {
@@ -179,8 +181,7 @@ class HumanReadableServiceTest
         .thenReturn(hrText);
 
     String generatedHumanReadable =
-        humanReadableService.generateMeasureHumanReadable(
-            madieMeasure, bundle, effectiveDataRequirements);
+        humanReadableService.generateMeasureHumanReadable(madieMeasure, bundle);
     assertNotNull(generatedHumanReadable);
     assertTrue(generatedHumanReadable.contains(hrText));
   }
@@ -200,8 +201,7 @@ class HumanReadableServiceTest
     le.setIncludeResolver(this);
     var hr = new HumanReadableService(le);
 
-    var generatedHumanReadable =
-        hr.generateMeasureHumanReadable(madieMeasure, bundle, effectiveDataRequirements);
+    var generatedHumanReadable = hr.generateMeasureHumanReadable(madieMeasure, bundle);
     assertNotNull(generatedHumanReadable);
   }
 
@@ -222,22 +222,12 @@ class HumanReadableServiceTest
 
     when(liquidEngine.evaluate(
             any(LiquidEngine.LiquidDocument.class),
-            argThat(
-                (measure) -> {
-                  org.hl7.fhir.r5.model.Measure m = (org.hl7.fhir.r5.model.Measure) measure;
-
-                  org.hl7.fhir.r5.model.Library library =
-                      (org.hl7.fhir.r5.model.Library) m.getContained().get(0);
-                  ParameterDefinition paramDef = library.getParameter().get(0);
-
-                  return "Period".equals(paramDef.getType().toCode());
-                }),
+            any(org.hl7.fhir.r5.model.Measure.class),
             any()))
         .thenReturn(hrText);
 
     String generatedHumanReadable =
-        humanReadableService.generateMeasureHumanReadable(
-            madieMeasure, bundle, effectiveDataRequirements);
+        humanReadableService.generateMeasureHumanReadable(madieMeasure, bundle);
     assertNotNull(generatedHumanReadable);
     assertTrue(generatedHumanReadable.contains(hrText));
   }
@@ -246,7 +236,7 @@ class HumanReadableServiceTest
   public void generateHumanReadableThrowsResourceNotFoundExceptionForNoBundle() {
     assertThrows(
         ResourceNotFoundException.class,
-        () -> humanReadableService.generateMeasureHumanReadable(madieMeasure, null, null));
+        () -> humanReadableService.generateMeasureHumanReadable(madieMeasure, null));
   }
 
   @Test
@@ -258,9 +248,7 @@ class HumanReadableServiceTest
 
     assertThrows(
         ResourceNotFoundException.class,
-        () ->
-            humanReadableService.generateMeasureHumanReadable(
-                madieMeasure, bundle, effectiveDataRequirements));
+        () -> humanReadableService.generateMeasureHumanReadable(madieMeasure, bundle));
   }
 
   @Test
@@ -284,9 +272,7 @@ class HumanReadableServiceTest
 
     assertThrows(
         HumanReadableGenerationException.class,
-        () ->
-            humanReadableService.generateMeasureHumanReadable(
-                madieMeasure, bundle, effectiveDataRequirements));
+        () -> humanReadableService.generateMeasureHumanReadable(madieMeasure, bundle));
   }
 
   @Test
